Set theme when opening the app from a push notification
